Added improved data quality metrics, new plot settings toolbar icon
Previous data quality metric was only based on percentage CV50 was to perfect reproducibility, but now the reproducibility metric is based on CV AUC normalized to the maximum possible AUC.
Skewness score is now the area between the mean and median CV curves normalized to CV AUC
overall quality is the reproducibility score multiplied by (100 - skewness score)

Skewness should be as a metric for how how consistent detection of feature is, e.g. having a lot of features that can only be detected in one or two replicates will cause show up as a high skewness value.